Initial commit for API (#7)
* Initial commit

First commit containing code for the silero-rust api. Adds vad_session and vad_environment files. Commented out code in silero_example to test if it works without environment building.

* Response to comments

Update in response to comments, most notably added a check on model location to see if supplied path is valid

* Response to more comments

Responded to more comments left on the code review most noteworthy being adding an additional method to for wav files to be processed rejecting any with an incompatible sample rate

* Added tests for run_vad_wav