Threshold and return type configureable
Made threshold and return type configureable. Added in a struct for return values that has start time, end time, probability and a boolean for if the threshold lower than the probability for a window. A new struct called audio window is included that contains information on the start an end time of a window.