8 make threshold and return type configureable (#14)
* Added functions requested in ticket

Added the ability to change voice activation threshold and whether probabilities or booleans are returned.   Also added method that can return a result straight from a wav file rather than needing the audio data already extracted.

* Threshold and return type configureable

Made threshold and return type configureable. Added in a struct for return values that has start time, end time, probability and a boolean for if the threshold lower than the probability for a window. A new struct called audio window is included that contains information on the start an end time of a window.

* Renaming

Some renaming to improve readability